requirements.txt
absl-py==1.4.0
appdirs==1.4.4
astunparse==1.6.3
audiomentations==0.33.0
audioread==3.0.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
contourpy==1.0.7
cycler==0.11.0
Cython==0.29.33
decorator==5.1.1
filelock==3.12.2
flatbuffers==23.1.21
fonttools==4.38.0
fsspec==2023.6.0
gast==0.4.0
google-auth==2.16.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.1
h5py==3.8.0
hdbscan==0.8.29
huggingface-hub==0.16.4
idna==3.4
importlib-metadata==6.0.0
joblib==1.2.0
julius==0.2.7
keras==2.11.0
kiwisolver==1.4.4
libclang==15.0.6.1
librosa==0.9.2
llvmlite==0.39.1
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.6.3
nnAudio==0.3.2
numba==0.56.4
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
pooch==1.6.0
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydub==0.25.1
pynndescent==0.5.8
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
PyWavelets==1.4.1
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
resampy==0.2.2
rsa==4.9
safetensors==0.3.1
scikit-learn==1.2.1
scipy==1.10.0
seaborn==0.12.2
simclr==1.0.2
six==1.16.0
soundfile==0.11.0
soxr==0.3.6
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.30.0
tensorflow-macos==2.11.0
termcolor==2.2.0
threadpoolctl==3.1.0
timm==0.9.2
torch==1.13.1
torchaudio==0.13.1
torchopenl3==1.0.1
torchsummary==1.5.1
torchvision==0.14.1
tqdm==4.64.1
typing_extensions==4.4.0
umap-learn==0.5.3
urllib3==1.26.14
Werkzeug==2.2.2
wrapt==1.14.1
yacs==0.1.8
zipp==3.12.0
